Convert some absolute paths to relative to project root
This allows execution on other computers than ruil's. We just strip
/home/ruil/workspace/git/helloqa/ from many paths. This may require
adding a setting in Eclipse to set the working directory if you
aren't using the commandline execution.

There are still some absolute paths, but they lead to files that
are not included in Git and with no instructions to acquire them,
so we'll fix the code to run without these altogether.